Create README.md
Participants are encouraged to innovate by developing their unique functions (f(x)) to tackle the given challenge.
Participants have the liberty to utilize any tech stack of their preference for model development. This flexibility allows them to harness the tools and technologies they are most adept at, facilitating the creation of effective and inventive solutions and deriving the mathematical function for this Hackathon.